app/api/requirements.txt
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
beanie==1.26.0
beautifulsoup4==4.12.3
bs4==0.0.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.7.1
detect==2020.12.3
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
langdetect==1.0.9
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.0
multidict==6.0.5
numpy==2.0.0
orjson==3.10.6
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tinycss2==1.3.0
toml==0.10.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
validators==0.28.3
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
